#165:storybookのテストを更新
